+++ pytsite/auth_ui/_model.py
@@ -34,11 +34,11 @@ class UserUI(_auth.model.User, _odm_ui.UIMixin):
         """Hook.
         """
         if field_name == 'profile_view_url':
-            value = _router.ep_url('pytsite.auth_ui.eps.profile_view', {'uid': self.id})
+            value = _router.ep_url('pytsite.auth_ui.eps.profile_view', {'nickname': self.nickname})
 
         if field_name == 'profile_edit_url':
             value = _router.ep_url('pytsite.auth_ui.eps.profile_edit', {
-                'uid': self.id,
+                'nickname': self.nickname,
                 '__form_redirect': _router.current_url(),
             })