+++ pytsite/content/_widget.py
@@ -1,7 +1,7 @@
 """Content Widgets.
 """
 from pytsite import taxonomy as _taxonomy, auth as _auth, widget as _widget, html as _html, lang as _lang, \
-    router as _router
+    router as _router, tpl as _tpl
 from . import _model, _functions
 
 __author__ = 'Alexander Shepetko'
@@ -44,21 +44,36 @@ class EntityTagCloud(_taxonomy.widget.Cloud):
         return self._entity.tags
 
 
-class SearchInput(_widget.Base):
+class Search(_widget.Base):
     def __init__(self, model: str, **kwargs):
+        """Init.
+        """
         super().__init__(**kwargs)
+        self._value = _router.request.values_dict.get('search', '')
         self._model = model
+        self._title_tag = kwargs.get('title_tag', 'h3')
 
-    def render(self) -> _html.Element:
-        value = _router.request.values_dict.get('search', '')
-        form = _html.Form(cls='wrapper form-inline', method='GET')
-        form.append(_html.Input(type='text', cls='form-control', name='search',  required=True,
-                                placeholder=_lang.t('pytsite.content@search_input_placeholder'),
-                                value=value))
-        form.set_attr('action', _router.ep_url('pytsite.content.ep.index', {
-            'model': self._model,
-        }))
+        self._form = _html.Form(cls='wrapper form-inline', method='GET')
+        self._form.append(_html.Input(type='text', cls='form-control', name='search',  required=True, value=self.value,
+                          placeholder=_lang.t('pytsite.content@search_input_placeholder')))
+        self._form.set_attr('action', _router.ep_url('pytsite.content.ep.search', {'model': self._model}))
 
         btn = _html.Button(type='submit', cls='btn btn-default')
-        form.append(btn.append(_html.I(cls='fa fa-search')))
-        return self._group_wrap(form, False)
+        self._form.append(btn.append(_html.I(cls='fa fa-search')))
+
+    @property
+    def title_tag(self) -> str:
+        return self._title_tag
+
+    @property
+    def model(self) -> str:
+        return self._model
+
+    @property
+    def form(self) -> _html.Element:
+        return self._form
+
+    def render(self) -> _html.Element:
+        """Render the widget.
+        """
+        return self._group_wrap(_tpl.render('pytsite.content@widget/search', {'widget': self}))